refactor(issue-130): split PageDef into PageSpec + TransitionSpec
Lands candidate 8 from docs/architecture-deepening.md (grilled
2026-05-14). The old physics/PageDef.ts mixed two authoring intents
into one grab-bag interface: route-author (`gravity`, `cards`,
`sections?`) and transition-author (`transitions?`, `siblingOrder?`).
Every consumer had to import the union even though physics never
reads `transitions?` and dispatch never reads `cards`.

Type split:

- web/src/physics/PageSpec.ts — new home for the route-author half
  (`PageSpec` interface + `Cardinal`, `CardKind`, `CardSpec`,
  `Buoyancy`, `ParentRef`, `SectionsConfig`, `AuthorSectionDef`,
  `ChainNavTarget`, `buoyancyForKind`). The "what the route is" half.
- web/src/transitions/TransitionSpec.ts — new home for the
  transition-author half (`TransitionSpec` interface +
  `TransitionId`). Imports nothing from physics, completing the
  no-backward-cross-folder-type-dep rule.
- web/src/routes/PageDef.ts — re-composes them:
  `PageDef = PageSpec & TransitionSpec`. Routes author the
  intersection; subsystems read only the half they need.

Freebies folded in (per the grill):

- web/src/transitions/pageDefs.ts → web/src/routes/pageDefs.ts.
  The static route-keyed map is route-author intent, not transition
  machinery. Importers updated.
- `sectionsPushHistory?` field deleted (zero production consumers).
- Stale 21b/pattern-matching doc-comments removed from
  routes/pageDefs.ts and transitions/edges.ts (21b shipped the
  runtime registry instead).

Forced-by-design tightenings beyond the issue's stated importer list:

- usePageDef, PhysicsPage, PhysicsCard, partitionPageDef — all
  physics/layout consumers that previously imported PageDef now
  take PageSpec. PageDef ⊆ PageSpec structurally so callers don't
  change. This is forced because physics/ and layout/ may not
  import from routes/.

CONTEXT.md: PageDef entry rewritten; new PageSpec and TransitionSpec
entries added.

Dispatch test seam-demonstration: a new test in
transitions/dispatch.test.ts constructs a TransitionSpec literal
directly and spreads it into a stubbed PageDef, demonstrating that
dispatch's contract is over the TransitionSpec half for non-section
transitions.

Closes #130
