+++ web/src/routes/PageDef.ts
@@ -1,0 +1,4 @@
+import type { PageSpec } from '../physics/PageSpec'
+import type { TransitionSpec } from '../transitions/TransitionSpec'
+
+export type PageDef = PageSpec & TransitionSpec
